requirements.txt
playwright==1.58.0
pytest==9.0.2
pytest-playwright==0.7.2
pytest-html==4.2.0
pytest-xdist==3.5.0
python-dotenv==1.2.1
